--- a/dbms/src/Storages/Transaction/ReadIndexWorker.cpp
+++ b/dbms/src/Storages/Transaction/ReadIndexWorker.cpp
@@ -20,6 +20,7 @@
 #include <fmt/chrono.h>
 
 #include <memory>
+#include <queue>
 
 namespace DB
 {